Initial Agent Gallery page working with data being pulled in. Need to clean up styling
