+++ client/src/assets/agents.json
@@ -1,7 +1,7 @@
 [
   {
     "id": 1,
-    "name": "Empress",
+    "name": "Chad",
     "avatar": "agent1.jpg",
     "role": "Nature's Guardian",
     "shortDescription": "...",
