+++ client/src/hooks/useCharacters.ts
@@ -22,6 +22,7 @@ const useCharacters = () => {
     if (!selectedCharacter && processedCharacters.length > 0) {
       handleCharacterSelect(processedCharacters[0]); // Select the first character by default.
     }
+    console.log('[ProcessedCharacters] - processedCharacters:', processedCharacters);  
   }, [processedCharacters, selectedCharacter, handleCharacterSelect]);
 
   return {
