+++ client/src/pages/AgentGallery.tsx
@@ -12,20 +12,24 @@ import image2 from '../assets/agent-images/agent2.jpg';
 import image3 from '../assets/agent-images/agent3.jpg';
 import image4 from '../assets/agent-images/agent4.jpg';
 import agentsData from '../assets/agents.json'; // Import the JSON file
+import { Agent } from '../interfaces/AgentInterfaces';
+import useCharacters from '../hooks/useCharacters';
+import RandomAgentCard from '../components/RandomAgentCard'; // Import the new component
+import LoadedAgentCard from '../components/LoadedAgentCard'; // Import the new component
 
-// Define the Agent type
-interface Agent {
-  id: number;
-  name: string;
-  avatar: string;
-  role: string;
-  shortDescription?: string; // Optional property
-  tags: string[];
-  personality: string;
-  communicationStyle: string;
-  abilities: string[];
-  placeholder?: boolean; // Add optional placeholder property
-}
+// // Define the Agent type
+// interface Agent {
+//   id: number;
+//   name: string;
+//   avatar: string;
+//   role: string;
+//   shortDescription?: string; // Optional property
+//   tags: string[];
+//   personality: string;
+//   communicationStyle: string;
+//   abilities: string[];
+//   placeholder?: boolean; // Add optional placeholder property
+// }
 
 // Define the props for AgentCard
 interface AgentCardProps {
@@ -35,6 +39,7 @@ interface AgentCardProps {
   isUserAgent?: boolean; // Flag to indicate if it's a user agent
   setRandomAgents: React.Dispatch<React.SetStateAction<Agent[]>>; // Add setRandomAgents prop
   generateRandomAgent: () => Agent; // Add generateRandomAgent prop
+  isLoadedAgent?: boolean; // New prop to indicate if the agent is loaded
 }
 
 // Regenerate Button Component
@@ -59,190 +64,9 @@ const RegenerateButton: React.FC<RegenerateButtonProps> = ({
   );
 };
 
-const AgentCard: React.FC<AgentCardProps> = ({
-  agent,
-  onSelect,
-  onAdd,
-  isUserAgent,
-  setRandomAgents,
-  generateRandomAgent,
-}) => {
-  const [isFlipped, setIsFlipped] = useState(false);
-  const [isRegenerating, setIsRegenerating] = useState(false); // State to track regeneration
-
-  const handleRegenerate = () => {
-    setIsRegenerating(true); // Show placeholder
-
-    // Simulate regeneration delay (remove in actual implementation)
-    setTimeout(() => {
-      setRandomAgents((prevAgents: Agent[]) =>
-        prevAgents.map((a: Agent) =>
-          a.id === agent.id ? generateRandomAgent() : a
-        )
-      );
-      setIsRegenerating(false); // Hide placeholder
-    }, 500); // Adjust delay as needed
-  };
-
-  return (
-    <div className="relative">
-      {agent.placeholder ? (
-        <div className="absolute inset-0 flex items-center justify-center bg-gray-800/75">
-          <p className="text-white">Regenerating...</p>
-        </div>
-      ) : (
-        <div
-          className="perspective w-64 h-[500px]"
-          onMouseEnter={() => setIsFlipped(true)}
-          onMouseLeave={() => setIsFlipped(false)}
-          onClick={(e) => {
-            e.stopPropagation();
-            onSelect(agent);
-          }}
-        >
-          <div
-            className={`relative w-full h-full duration-500 preserve-3d ${
-              isFlipped ? 'rotate-y-180' : ''
-            }`}
-          >
-            {/* Front of card */}
-            <div className="absolute w-full h-full backface-hidden">
-              <div className="w-full h-full bg-gray-800 rounded-lg overflow-hidden shadow-xl border border-purple-500/30">
-                {/* Image container - 80% of card height */}
-                <div className="relative h-[400px]">
-                  <img
-                    src={agent.avatar}
-                    alt={agent.name}
-                    className="w-full h-full object-cover"
-                  />
-                  <div className="absolute bottom-0 left-0 right-0 bg-gradient-to-t from-gray-900 to-transparent h-16" />
-                </div>
-
-                {/* Title area - 20% of card height */}
-                <div className="h-[100px] p-4 bg-gray-800/95">
-                  <h3 className="text-xl font-bold text-gray-100 mb-1">
-                    {agent.name}
-                  </h3>
-                  <p className="text-purple-300 text-sm">{agent.role}</p>
-                </div>
-              </div>
-            </div>
-
-            {/* Back of card */}
-            <div className="absolute w-full h-full backface-hidden rotate-y-180">
-              <div className="w-full h-full bg-gray-800 rounded-lg p-4 shadow-xl border border-purple-500/30">
-                {/* Header with small image */}
-                <div className="flex gap-4 mb-4">
-                  <img
-                    src={agent.avatar}
-                    alt={agent.name}
-                    className="w-20 h-20 rounded-lg object-cover"
-                  />
-                  <div>
-                    <h3 className="text-xl font-bold text-gray-100">
-                      {agent.name}
-                    </h3>
-                    <p className="text-purple-400 text-sm">{agent.role}</p>
-                  </div>
-                </div>
-
-                {/* Content sections */}
-                <div className="space-y-4 overflow-auto max-h-[350px] pr-2">
-                  <div>
-                    <div className="flex items-center gap-2 text-gray-300 mb-1">
-                      <Heart className="w-4 h-4 text-purple-400" />
-                      <span className="font-medium">Personality</span>
-                    </div>
-                    <p className="text-gray-400 text-sm">{agent.personality}</p>
-                  </div>
-
-                  <div>
-                    <div className="flex items-center gap-2 text-gray-300 mb-1">
-                      <MessageCircle className="w-4 h-4 text-purple-400" />
-                      <span className="font-medium">Communication</span>
-                    </div>
-                    <p className="text-gray-400 text-sm">
-                      {agent.communicationStyle}
-                    </p>
-                  </div>
-
-                  <div>
-                    <div className="flex items-center gap-2 text-gray-300 mb-1">
-                      <Sparkles className="w-4 h-4 text-purple-400" />
-                      <span className="font-medium">Abilities</span>
-                    </div>
-                    <ul className="text-gray-400 text-sm pl-4 list-disc space-y-1">
-                      {agent.abilities.map((ability, index) => (
-                        <li key={index}>{ability}</li>
-                      ))}
-                    </ul>
-                  </div>
-                </div>
-
-                {/* Tags at bottom */}
-                <div className="absolute bottom-12 left-4 right-4">
-                  <div className="flex gap-2 flex-wrap">
-                    {agent.tags.map((tag, index) => (
-                      <span
-                        key={index}
-                        className="px-2 py-1 bg-purple-900/50 rounded-full text-xs text-purple-300"
-                      >
-                        {tag}
-                      </span>
-                    ))}
-                  </div>
-                </div>
-
-                {/* Action button at bottom */}
-                <div className="absolute bottom-2 left-4 right-4">
-                  {isUserAgent ? (
-                    <button
-                      className="w-full px-4 py-2 bg-purple-600 rounded-md hover:bg-purple-700 flex items-center justify-center gap-2"
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        onSelect(agent);
-                      }}
-                    >
-                      <CheckCircle className="w-4 h-4" />
-                      Select Agent
-                    </button>
-                  ) : (
-                    <button
-                      className="w-full px-4 py-2 bg-blue-600 rounded-md hover:bg-blue-700 flex items-center justify-center gap-2"
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        onAdd && onAdd(agent);
-                      }}
-                    >
-                      <PlusCircle className="w-4 h-4" />
-                      Add Agent
-                    </button>
-                  )}
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Regenerate button below the card */}
-      {!isUserAgent && (
-        <div className="mt-2 w-64 mx-auto">
-          <RegenerateButton onRegenerate={handleRegenerate} />
-        </div>
-      )}
-
-      {/* Placeholder while regenerating */}
-      {isRegenerating && (
-        <div className="absolute inset-0 flex items-center justify-center bg-gray-800/75">
-          <p className="text-white">Regenerating...</p>
-        </div>
-      )}
-    </div>
-  );
-};
-
 const AgentGallery: React.FC = () => {
+  const { characters: loadedAgents, loading, error } = useCharacters();
+  console.log('[AgentGallery] - loaded agents:', loadedAgents);
   const [selectedAgent, setSelectedAgent] = useState<Agent | null>(null);
   const [filter, setFilter] = useState('all'); // 'all', 'random', or 'yourAgents'
   const [randomAgents, setRandomAgents] = useState<Agent[]>([]);
@@ -361,12 +185,12 @@ const AgentGallery: React.FC = () => {
 
   useEffect(() => {
     // Load agents from JSON
-    const randomAgents = [...Array(3)].map(() => {
+    const randomAgents = [...Array(3)].map((_, index) => {
       const randomIndex = Math.floor(Math.random() * agentsData.length);
       const agent = agentsData[randomIndex];
       return {
         ...agent,
-        id: Date.now() + randomIndex,
+        id: Date.now() + index, // Ensure unique IDs by adding index
         avatar: agentImages[Math.floor(Math.random() * agentImages.length)],
       }; // Ensure unique IDs
     });
@@ -375,7 +199,7 @@ const AgentGallery: React.FC = () => {
       const agent = agentsData[index % agentsData.length]; // Cycle through agents
       return {
         ...agent,
-        id: Date.now() + index + 1000,
+        id: Date.now() + index + 1000, // Ensure unique IDs by adding index
         avatar: agentImages[Math.floor(Math.random() * agentImages.length)],
       }; // Ensure unique IDs
     });
@@ -433,7 +257,7 @@ const AgentGallery: React.FC = () => {
               <h2 className="text-xl font-semibold mb-4">Random Agents</h2>
               <div className="flex flex-wrap gap-6 justify-center">
                 {randomAgents.map((agent) => (
-                  <AgentCard
+                  <RandomAgentCard
                     key={agent.id}
                     agent={agent}
                     onSelect={setSelectedAgent}
@@ -441,21 +265,23 @@ const AgentGallery: React.FC = () => {
                     isUserAgent={false}
                     setRandomAgents={setRandomAgents}
                     generateRandomAgent={generateRandomAgent}
+                    isLoadedAgent={true}
                   />
                 ))}
               </div>
               <h2 className="text-xl font-semibold mt-8 mb-4">Your Agents</h2>
               <div className="flex flex-wrap gap-6 justify-center">
-                {yourAgents.map((agent) => (
-                  <AgentCard
-                    key={agent.id}
-                    agent={agent}
-                    onSelect={setSelectedAgent}
-                    onAdd={handleAddAgent}
-                    isUserAgent={true}
-                    setRandomAgents={setRandomAgents}
-                    generateRandomAgent={generateRandomAgent}
-                  />
+                {loadedAgents.map((agent) => (
+                   <LoadedAgentCard
+                  //  key={agent.id}
+                   agent={agent}
+                   onSelect={setSelectedAgent}
+                  //  onAdd={handleAddAgent}
+                  //  isUserAgent={false}
+                  //  setRandomAgents={setRandomAgents}
+                  //  generateRandomAgent={generateRandomAgent}
+                  //  isLoadedAgent={true}
+                 />
                 ))}
               </div>
             </>
@@ -467,7 +293,7 @@ const AgentGallery: React.FC = () => {
               <h2 className="text-xl font-semibold mb-4">Random Agents</h2>
               <div className="flex flex-wrap gap-6 justify-center">
                 {randomAgents.map((agent) => (
-                  <AgentCard
+                  <RandomAgentCard
                     key={agent.id}
                     agent={agent}
                     onSelect={setSelectedAgent}
@@ -475,6 +301,7 @@ const AgentGallery: React.FC = () => {
                     isUserAgent={false}
                     setRandomAgents={setRandomAgents}
                     generateRandomAgent={generateRandomAgent}
+                    isLoadedAgent={true}
                   />
                 ))}
               </div>
@@ -486,16 +313,12 @@ const AgentGallery: React.FC = () => {
             <>
               <h2 className="text-xl font-semibold mb-4">Your Agents</h2>
               <div className="flex flex-wrap gap-6 justify-center">
-                {yourAgents.map((agent) => (
-                  <AgentCard
-                    key={agent.id}
-                    agent={agent}
-                    onSelect={setSelectedAgent}
-                    onAdd={handleAddAgent}
-                    isUserAgent={true}
-                    setRandomAgents={setRandomAgents}
-                    generateRandomAgent={generateRandomAgent}
-                  />
+                {loadedAgents.map((agent) => (
+          <LoadedAgentCard
+          key={agent.id}
+          agent={agent}
+          onSelect={setSelectedAgent}
+        />
                 ))}
               </div>
             </>
